update password fix
